+++ packages/awino-ui/src/components/pages/earn/deposit/AssetSection/AssetSection.tsx
@@ -2,8 +2,6 @@ import { useMemo, useState, useEffect, useCallback } from 'react';
 
 import { useRouter } from 'next/router';
 
-import { Box } from '@mui/material';
-import { styled } from '@mui/material/styles';
 import { GridEventListener, GridEvents, GridRowsProp, GridSortModel } from '@mui/x-data-grid';
 
 import { TABLE_ROWS_PER_PAGE, TABLE_ROWS_PER_PAGE_OPTIONS } from '@/app/constants';
@@ -13,6 +11,7 @@ import loadData from '@/app/data';
 import DataGrid from '@/components/general/DataGrid/DataGrid';
 import GridPagination from '@/components/general/GridPagination/GridPagination';
 import Label from '@/components/general/Label/Label';
+import Panel from '@/components/general/Panel/Panel';
 import Search from '@/components/general/Search/Search';
 import Switch from '@/components/general/Switch/Switch';
 import Section from '@/components/layout/Section/Section';
@@ -20,43 +19,6 @@ import usePageTranslation from '@/hooks/usePageTranslation';
 import { RowsState } from '@/types/app';
 
 import getColumns from './columns';
-
-const Panel = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  borderRadius: +theme.shape.borderRadius * 5,
-  backgroundColor: theme.palette.background.transparent,
-  '.header': {
-    display: 'flex',
-    flexWrap: 'wrap',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    gap: theme.spacing(10),
-    padding: theme.spacing(5.5, 6.5, 5),
-    margin: theme.spacing(0, 0, 15, 0),
-    borderRadius: +theme.shape.borderRadius * 5,
-    backgroundColor: theme.palette.background.transparent,
-    '.aside': {
-      display: 'flex',
-      alignItems: 'center',
-    },
-  },
-  '.content': {
-    padding: theme.spacing(4, 12.5, 10),
-    '.table-container': {
-      height: 888 /* 66 * 10 + 12 * 10 - 12 */,
-      width: '100%',
-    },
-  },
-  [theme.breakpoints.up('md')]: {
-    '.header': {
-      display: 'flex',
-      alignItems: 'center',
-      gap: theme.spacing(18),
-      padding: theme.spacing(5.5, 12.5, 5),
-    },
-  },
-}));
 
 // interface Props {
 //   total: TotalAssetSize;